+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.2.0
 
+### Enhancements
+- Domains can now be imported using `spout import_domain CSVFILE`
+
 ## 0.1.0 (May 21, 2013)
 
 ### Enhancements